get quote functionality
get_quote method in class LoginScreenSucess. Method executes when the enlighten me button gets pressed. Method gets the list of txt files in quotes folder, then picks the file corresponding with the user's feeling. It then reads the file, and randomly picks a line (quote) and displays it in the quote label. If the users feeling is not supported, a message appears saying "try another feeling"